// src/interfaces/Producto.ts
// Generated by https://quicktype.io

export interface Producto {
  id: number;
  idCategoria: number;
  nombreCategoria: string;
  idTienda: number;
  nombreTienda: string;
  nombreProducto: string;
  presentacion: string;
  precio: number;
  enOferta: boolean;
  porcentajeDescuento: number;
  precioOferta: number;
  fechaLimiteOferta: null;
  imagenes: Imagenes[];
}

export interface Imagenes {
  id: number;
  idProducto: number;
  nombreProducto: string;
  urlImagenProducto: string;
}

// Generated by https://quicktype.io

export interface ProductoFavoritoRequest {
  userId: string;
  idProducto: number;
}